영양 정보 호출 완료

+++ src/main/java/klj/project/service/user/baord/BoardService.java
@@ -44,7 +44,7 @@ public class BoardService {
 		return boardfindList;
 	}
 
-	public List<NutritionResDto> getNutriInfo(int boardId) {
+	public List<NutritionResDto> getNutriInfo(Long boardId) {
 		List<NutritionResDto> nutriInfo = boardQuerydslRepository.getNutriInfo(boardId);
 		
 		return nutriInfo;
